feat: reanme table
