+++ dms-api/src/main/resources/static/i18n/messages_en_US.properties
@@ -30,6 +30,8 @@ response.error.duplicate.data=Duplicate data
 response.error.datasource.connection=Database connection exception
 response.error.datasource.catalog=Failed to get database information, please check the database connection information
 response.error.file.catalog.notNull=Directory is not empty
+response.error.db.typeNotSupported=type not supported
+
 # dms privileges
 # module privilege
 dms.p.sys.dft.dft.show=System Management
