+++ dms-api/src/main/resources/static/i18n/messages_zh_CN.properties
@@ -30,6 +30,7 @@ response.error.duplicate.data=\u6570\u636E\u91CD\u590D
 response.error.datasource.connection=\u6570\u636E\u5E93\u8FDE\u63A5\u5F02\u5E38
 response.error.datasource.catalog=\u83B7\u53D6\u6570\u636E\u5E93\u4FE1\u606F\u5931\u8D25\uFF0C\u8BF7\u68C0\u67E5\u6570\u636E\u5E93\u8FDE\u63A5\u4FE1\u606F
 response.error.file.catalog.notNull=\u76EE\u5F55\u4E0D\u4E3A\u7A7A
+response.error.db.typeNotSupported=\u4E0D\u652F\u6301\u7684\u7C7B\u578B
 # dms privileges
 # module privilege
 dms.p.sys.dft.dft.show=\u7CFB\u7EDF\u7BA1\u7406
