+++ dms-plugins/dms-plugins-datasource/src/main/java/com/basedt/dms/plugins/datasource/FunctionHandler.java
@@ -15,31 +15,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.basedt.dms.common.exception;
 
-import com.basedt.dms.common.enums.ResponseCode;
-import com.basedt.dms.common.utils.I18nUtil;
+package com.basedt.dms.plugins.datasource;
 
-public class DmsException extends Exception {
+import com.basedt.dms.plugins.datasource.dto.FunctionDTO;
 
-    private String exceptionCode;
+import javax.sql.DataSource;
+import java.sql.SQLException;
+import java.util.List;
+import java.util.Map;
 
+public interface FunctionHandler {
 
-    public DmsException(String code, String message) {
-        super(message);
-        setExceptionCode(code);
-    }
+    void initialize(DataSource dataSource, Map<String, String> config);
 
-    public DmsException(ResponseCode responseCode) {
-        super(I18nUtil.get(responseCode.getLabel()));
-        setExceptionCode(responseCode.getValue());
-    }
+    List<FunctionDTO> listFunctions(String catalog, String schemaPattern, String functionPattern) throws SQLException;
 
-    public String getExceptionCode() {
-        return exceptionCode;
-    }
+    List<FunctionDTO> listFunctionDetails(String catalog, String schemaPattern, String functionPattern) throws SQLException;
 
-    public void setExceptionCode(String exceptionCode) {
-        this.exceptionCode = exceptionCode;
-    }
+    FunctionDTO getFunctionDetail(String catalog, String schemaPattern, String functionPattern) throws SQLException;
+
 }
