+++ dms-plugins/dms-plugins-datasource/src/main/java/com/basedt/dms/plugins/datasource/impl/mysql/MysqlFunctionHandler.java
@@ -1,0 +1,51 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.basedt.dms.plugins.datasource.impl.mysql;
+
+import cn.hutool.core.util.StrUtil;
+import com.basedt.dms.plugins.datasource.dto.FunctionDTO;
+import com.basedt.dms.plugins.datasource.impl.jdbc.JdbcFunctionHandler;
+
+import java.sql.SQLException;
+import java.util.List;
+
+public class MysqlFunctionHandler extends JdbcFunctionHandler {
+
+    @Override
+    public List<FunctionDTO> listFunctionDetails(String catalog, String schemaPattern, String functionPattern) throws SQLException {
+        String sql = "select " +
+                " null as catalog_name," +
+                " r.routine_schema as schema_name," +
+                " r.routine_name as object_name," +
+                " 'FUNCTION' as object_type," +
+                " r.routine_definition as source_code," +
+                " r.created as create_time," +
+                " r.last_altered as last_ddl_time," +
+                " r.routine_comment as remark" +
+                " from information_schema.routines r" +
+                " where 1=1 ";
+        if (StrUtil.isNotEmpty(schemaPattern)) {
+            sql += " and r.routine_schema = '" + schemaPattern + "'";
+        }
+        if (StrUtil.isNotEmpty(functionPattern)) {
+            sql += " and r.routine_name = '" + functionPattern + "'";
+        }
+        return super.listFunctionFromDB(sql);
+    }
+}
