+++ dms-plugins/dms-plugins-datasource/src/main/java/com/basedt/dms/plugins/datasource/impl/postgre/PostgreCatalogHandler.java
@@ -1,0 +1,68 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.basedt.dms.plugins.datasource.impl.postgre;
+
+import com.basedt.dms.plugins.datasource.dto.SchemaDTO;
+import com.basedt.dms.plugins.datasource.dto.TypeInfoDTO;
+import com.basedt.dms.plugins.datasource.impl.jdbc.JdbcCatalogHandler;
+
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+import static com.basedt.dms.plugins.datasource.enums.DbObjectType.*;
+
+public class PostgreCatalogHandler extends JdbcCatalogHandler {
+
+    @Override
+    public List<SchemaDTO> listSchemas(String catalog, String schemaPattern) throws SQLException {
+        List<SchemaDTO> list = super.listSchemas(catalog, schemaPattern);
+        return list.stream().filter(s -> {
+            if ("pg_catalog".equalsIgnoreCase(s.getSchemaName()) ||
+                    "information_schema".equalsIgnoreCase(s.getSchemaName())) {
+                return false;
+            } else {
+                return true;
+            }
+        }).collect(Collectors.toList());
+    }
+
+    @Override
+    public Map<String, TypeInfoDTO> listDataType() throws SQLException {
+        Map<String, TypeInfoDTO> map = super.listDataType();
+        map.entrySet().removeIf((entry) -> entry.getKey().startsWith("_"));
+        return map;
+    }
+
+    @Override
+    public List<String> listObjectTypes() throws SQLException {
+        List<String> list = new ArrayList<String>() {{
+            add(TABLE.name());
+            add(VIEW.name());
+            add(MATERIALIZED_VIEW.name());
+            add(FOREIGN_TABLE.name());
+            add(FUNCTION.name());
+            add(INDEX.name());
+            add(SEQUENCE.name());
+        }};
+        return list.stream().map(String::toLowerCase).collect(Collectors.toList());
+    }
+}
