+++ dms-plugins/dms-plugins-datasource/src/main/java/com/basedt/dms/plugins/datasource/impl/postgre/PostgreMaterializedViewHandler.java
@@ -1,0 +1,66 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.basedt.dms.plugins.datasource.impl.postgre;
+
+import cn.hutool.core.util.StrUtil;
+import com.basedt.dms.plugins.datasource.dto.MaterializedViewDTO;
+import com.basedt.dms.plugins.datasource.impl.jdbc.JdbcMaterializedViewHandler;
+
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.basedt.dms.plugins.datasource.enums.DbObjectType.MATERIALIZED_VIEW;
+
+public class PostgreMaterializedViewHandler extends JdbcMaterializedViewHandler {
+
+    @Override
+    public List<MaterializedViewDTO> listMViewDetails(String catalog, String schemaPattern, String mViewPattern) throws SQLException {
+        List<MaterializedViewDTO> viewList = new ArrayList<>();
+        String sql = " select " +
+                " null as catalog_name," +
+                " n.nspname as schema_name," +
+                " c.relname as object_name," +
+                " 'MATERIALIZED_VIEW'as object_type," +
+                " d.description as remark," +
+                " v.definition as query_sql," +
+                " pg_table_size(concat_ws('.',n.nspname, c.relname)) as data_bytes," +
+                " null as create_time," +
+                " null as last_ddl_time" +
+                " from pg_catalog.pg_namespace n" +
+                " join pg_catalog.pg_class c " +
+                " on n.oid = c.relnamespace  " +
+                " join pg_catalog.pg_matviews v " +
+                " on n.nspname = v.schemaname " +
+                " and c.relname = v.matviewname " +
+                " left join pg_catalog.pg_description d " +
+                " on c.oid  = d.objoid " +
+                " and d.objsubid = 0" +
+                " and d.classoid  = 'pg_class'::regclass" +
+                " where c.relkind in ('" + PostgreObjectTypeMapper.mapToOrigin(MATERIALIZED_VIEW) + "')";
+        if (StrUtil.isNotEmpty(schemaPattern)) {
+            sql += " and n.nspname = '" + schemaPattern + "'";
+        }
+        if (StrUtil.isNotEmpty(mViewPattern)) {
+            sql += " and c.relname = '" + mViewPattern + "'";
+        }
+        return super.listMViewFromDB(sql);
+
+    }
+}
