+++ dms-plugins/dms-plugins-datasource/src/main/java/com/basedt/dms/plugins/datasource/impl/postgre/PostgrePluginImpl.java
@@ -22,23 +22,19 @@ import com.basedt.dms.common.constant.Constants;
 import com.basedt.dms.common.utils.DateTimeUtil;
 import com.basedt.dms.plugins.core.PluginInfo;
 import com.basedt.dms.plugins.core.PluginType;
-import com.basedt.dms.plugins.datasource.AbstractDataSourcePlugin;
-import com.basedt.dms.plugins.datasource.DataSourcePlugin;
-import com.basedt.dms.plugins.datasource.ViewHandler;
-import com.basedt.dms.plugins.datasource.dto.*;
+import com.basedt.dms.plugins.datasource.*;
+import com.basedt.dms.plugins.datasource.dto.ColumnDTO;
 import com.basedt.dms.plugins.datasource.enums.DataSourceType;
-import com.basedt.dms.plugins.datasource.enums.DbObjectType;
-import com.basedt.dms.plugins.datasource.utils.JdbcUtil;
+import com.basedt.dms.plugins.datasource.impl.jdbc.JdbcForeignTableHandler;
 import com.google.auto.service.AutoService;
 
 import java.math.BigDecimal;
-import java.sql.Date;
 import java.sql.*;
 import java.text.ParseException;
-import java.util.*;
-import java.util.stream.Collectors;
-
-import static com.basedt.dms.plugins.datasource.enums.DbObjectType.*;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Properties;
 
 @AutoService(DataSourcePlugin.class)
 public class PostgrePluginImpl extends AbstractDataSourcePlugin {
@@ -69,334 +65,64 @@ public class PostgrePluginImpl extends AbstractDataSourcePlugin {
     }
 
     @Override
-    public ViewHandler getViewHandler() {
-        PostgreViewHandler handler = new PostgreViewHandler();
-        handler.initialize(getDataSource(),new HashMap<>());
+    public CatalogHandler getCatalogHandler() {
+        PostgreCatalogHandler handler = new PostgreCatalogHandler();
+        handler.initialize(getDataSource(), new HashMap<>(), getDatabaseName());
         return handler;
     }
 
     @Override
-    public List<SchemaDTO> listSchemas(String catalog, String schemaPattern) throws SQLException {
-        List<SchemaDTO> list = super.listSchemas(catalog, schemaPattern);
-        return list.stream().filter(s -> {
-            if ("pg_catalog".equalsIgnoreCase(s.getSchemaName()) ||
-                    "information_schema".equalsIgnoreCase(s.getSchemaName())) {
-                return false;
-            } else {
-                return true;
-            }
-        }).collect(Collectors.toList());
+    public TableHandler getTableHandler() {
+        PostgreTableHandler handler = new PostgreTableHandler();
+        handler.initialize(getDataSource(), new HashMap<>());
+        return handler;
     }
 
     @Override
-    public List<TableDTO> listTableDetails(String catalog, String schemaPattern, String tablePattern, DbObjectType type) throws SQLException {
-        String sql = "select " +
-                " null as catalog_name," +
-                " n.nspname as schema_name," +
-                " c.relname as object_name," +
-                " case c.relkind when 'r' then 'TABLE' when 'i' then 'INDEX' when 'v' then 'VIEW' when 'S' then 'SEQUENCE'" +
-                " when 'f' then 'FOREIGN_TABLE' when 'm' then 'MATERIALIZED_VIEW' else null end as object_type," +
-                " d.description as remark," +
-                " pg_table_size(concat_ws('.',n.nspname, c.relname)) as data_bytes," +
-                " c.reltuples as table_rows," +
-                " null as create_time," +
-                " null as last_ddl_time," +
-                " null as last_access_time" +
-                " from pg_catalog.pg_namespace n" +
-                " join pg_catalog.pg_class c " +
-                " on n.oid = c.relnamespace  " +
-                " left join pg_catalog.pg_description d " +
-                " on c.oid  = d.objoid " +
-                " and d.objsubid = 0" +
-                " and d.classoid  = 'pg_class'::regclass" +
-                " where c.relkind in ('" + PostgreObjectTypeMapper.mapToOrigin(type) + "')";
-        if (StrUtil.isNotEmpty(schemaPattern)) {
-            sql += " and n.nspname = '" + schemaPattern + "'";
-        }
-        if (StrUtil.isNotEmpty(tablePattern)) {
-            sql += " and c.relname like '%" + tablePattern + "%'";
-        }
-        return super.listTableDetails(sql);
-    }
-
-
-    @Override
-    public List<TableDTO> listForeignTables(String catalog, String schemaPattern, String tablePattern) throws SQLException {
-        return listTableDetails(catalog, schemaPattern, tablePattern, DbObjectType.FOREIGN_TABLE);
+    public ViewHandler getViewHandler() {
+        PostgreViewHandler handler = new PostgreViewHandler();
+        handler.initialize(getDataSource(), new HashMap<>());
+        return handler;
     }
 
     @Override
-    public List<IndexDTO> listIndexes(String catalog, String schemaPattern, String tableName) throws SQLException {
-        return listIndexDetails(catalog, schemaPattern, tableName);
+    public ForeignTableHandler getForeignTableHandler() {
+        JdbcForeignTableHandler handler = new JdbcForeignTableHandler();
+        handler.initialize(getDataSource(), new HashMap<>());
+        return handler;
     }
 
     @Override
-    public List<IndexDTO> listIndexDetails(String catalog, String schemaPattern, String tableName) throws SQLException {
-        String sql = " select" +
-                " null as catalog_name," +
-                " n.nspname as schema_name," +
-                " c.relname as object_name," +
-                " 'INDEX'as object_type," +
-                " i.tablename as table_name," +
-                " pg_relation_size(pgi.indexrelid::regclass) as index_bytes," +
-                " pgi.indisunique as is_uniqueness," +
-                " a.columns as columns," +
-                " null as index_type," +
-                " null as create_time," +
-                " null as last_ddl_time" +
-                " from pg_catalog.pg_namespace n" +
-                " join pg_catalog.pg_class c" +
-                " on n.oid = c.relnamespace" +
-                " join pg_catalog.pg_indexes i" +
-                " on n.nspname = i.schemaname" +
-                " and c.relname = i.indexname" +
-                " join pg_catalog.pg_index pgi" +
-                " on c.oid = pgi.indexrelid" +
-                " left join (" +
-                "   SELECT" +
-                "        n.nspname as schemaname," +
-                "        t.relname AS tablename," +
-                "        c.relname AS indexname," +
-                "        string_agg(a.attname,',')  AS columns" +
-                "    FROM" +
-                "        pg_catalog.pg_index i" +
-                "    JOIN" +
-                "        pg_catalog.pg_class c ON c.oid = i.indexrelid" +
-                "    join pg_catalog.pg_namespace n on c.relnamespace  = n.oid" +
-                "    JOIN" +
-                "        pg_catalog.pg_class t ON t.oid = i.indrelid" +
-                "    JOIN" +
-                "        pg_catalog.pg_attribute a ON a.attrelid = t.oid AND a.attnum = ANY(i.indkey)" +
-                "    group by n.nspname,t.relname,c.relname" +
-                " ) a" +
-                "                 on a.schemaname = n.nspname" +
-                "                and a.tablename = i.tablename" +
-                "                and a.indexname = c.relname" +
-                " where c.relkind in ('" + PostgreObjectTypeMapper.mapToOrigin(INDEX) + "')";
-        if (StrUtil.isNotEmpty(schemaPattern)) {
-            sql += " and n.nspname = '" + schemaPattern + "'";
-        }
-        if (StrUtil.isNotEmpty(tableName)) {
-            sql += " and i.tablename = '" + tableName + "'";
-        }
-        return super.listIndexDetails(sql);
+    public FunctionHandler getFunctionHandler() {
+        PostgreFunctionHandler handler = new PostgreFunctionHandler();
+        handler.initialize(getDataSource(), new HashMap<>());
+        return handler;
     }
 
     @Override
-    public List<MaterializedViewDTO> listMViews(String catalog, String schemaPattern, String mViewPattern) throws SQLException {
-        return listMViewDetails(catalog, schemaPattern, mViewPattern);
+    public IndexHandler getIndexHandler() {
+        PostgreIndexHandler handler = new PostgreIndexHandler();
+        handler.initialize(getDataSource(), new HashMap<>());
+        return handler;
     }
 
     @Override
-    public List<MaterializedViewDTO> listMViewDetails(String catalog, String schemaPattern, String mViewPattern) throws SQLException {
-        List<MaterializedViewDTO> viewList = new ArrayList<>();
-        String sql = " select " +
-                " null as catalog_name," +
-                " n.nspname as schema_name," +
-                " c.relname as object_name," +
-                " 'MATERIALIZED_VIEW'as object_type," +
-                " d.description as remark," +
-                " v.definition as query_sql," +
-                " pg_table_size(concat_ws('.',n.nspname, c.relname)) as data_bytes," +
-                " null as create_time," +
-                " null as last_ddl_time" +
-                " from pg_catalog.pg_namespace n" +
-                " join pg_catalog.pg_class c " +
-                " on n.oid = c.relnamespace  " +
-                " join pg_catalog.pg_matviews v " +
-                " on n.nspname = v.schemaname " +
-                " and c.relname = v.matviewname " +
-                " left join pg_catalog.pg_description d " +
-                " on c.oid  = d.objoid " +
-                " and d.objsubid = 0" +
-                " and d.classoid  = 'pg_class'::regclass" +
-                " where c.relkind in ('" + PostgreObjectTypeMapper.mapToOrigin(MATERIALIZED_VIEW) + "')";
-        if (StrUtil.isNotEmpty(schemaPattern)) {
-            sql += " and n.nspname = '" + schemaPattern + "'";
-        }
-        if (StrUtil.isNotEmpty(mViewPattern)) {
-            sql += " and c.relname like '%" + mViewPattern + "%'";
-        }
-        return super.listMViewDetails(sql);
+    public MaterializedViewHandler getMaterializedViewHandler() {
+        PostgreMaterializedViewHandler handler = new PostgreMaterializedViewHandler();
+        handler.initialize(getDataSource(), new HashMap<>());
+        return handler;
     }
 
     @Override
-    public List<SequenceDTO> listSequences(String catalog, String schemaPattern, String sequencePattern) throws SQLException {
-        return listSequenceDetails(catalog, schemaPattern, sequencePattern);
-    }
-
-    @Override
-    public List<SequenceDTO> listSequenceDetails(String catalog, String schemaPattern, String sequencePattern) throws SQLException {
-        String sql = " select " +
-                " null as catalog_name," +
-                " n.nspname as schema_name," +
-                " c.relname as object_name," +
-                " 'SEQUENCE'as object_type," +
-                " s.start_value as start_value," +
-                " s.minimum_value as min_value," +
-                " s.maximum_value as max_value," +
-                " s.increment as increment_by," +
-                " case when cycle_option = 'YES' then 1 else 0 end as is_cycle," +
-                " null as last_value," +
-                " null as create_time," +
-                " null as last_ddl_time" +
-                " from pg_catalog.pg_namespace n" +
-                " join pg_catalog.pg_class c " +
-                " on n.oid = c.relnamespace  " +
-                " join information_schema.sequences s" +
-                " on n.nspname = s.sequence_schema " +
-                " and c.relname = s.sequence_name " +
-                " where c.relkind in ('" + PostgreObjectTypeMapper.mapToOrigin(SEQUENCE) + "')";
-        if (StrUtil.isNotEmpty(schemaPattern)) {
-            sql += " and n.nspname = '" + schemaPattern + "'";
-        }
-        if (StrUtil.isNotEmpty(sequencePattern)) {
-            sql += " and c.relname like '%" + sequencePattern + "%'";
-        }
-        return super.listSequenceDetails(sql);
-    }
-
-    @Override
-    public List<FunctionDTO> listFunctions(String catalog, String schemaPattern, String functionPattern) throws SQLException {
-        return listFunctionDetails(catalog, schemaPattern, functionPattern);
-    }
-
-    @Override
-    public List<FunctionDTO> listFunctionDetails(String catalog, String schemaPattern, String functionPattern) throws SQLException {
-        String sql = "select" +
-                "    catalog_name," +
-                "    schema_name," +
-                "    function_name as object_name," +
-                "    object_type," +
-                "    source_code," +
-                "    null as create_time," +
-                "    null as last_ddl_time" +
-                " from" +
-                "    (" +
-                "        select" +
-                "            null as catalog_name," +
-                "            n.nspname as schema_name," +
-                "            p.proname as function_name," +
-                "            'FUNCTION' as object_type," +
-                "            p.prosrc as source_code," +
-                "            row_number() over (partition by n.nspname,p.proname order by p.oid) as rn" +
-                "        from pg_catalog.pg_namespace n" +
-                "            join pg_catalog.pg_proc p" +
-                "            on p.pronamespace = n.oid" +
-                "        where p.prokind = 'f'" +
-                "    ) tt" +
-                " where tt.rn = 1 ";
-        if (StrUtil.isNotEmpty(schemaPattern)) {
-            sql += " and tt.schema_name = '" + schemaPattern + "'";
-        }
-        if (StrUtil.isNotEmpty(functionPattern)) {
-            sql += " and tt.function_name like '%" + functionPattern + "%'";
-        }
-        return super.listFunctionDetails(sql);
-    }
-
-    @Override
-    public List<ObjectDTO> listPkByTable(String catalog, String schemaPattern, String tableName) throws SQLException {
-        return getConstraint(catalog, schemaPattern, tableName, PK);
-    }
-
-    @Override
-    public List<ObjectDTO> listFkByTable(String catalog, String schemaPattern, String tableName) throws SQLException {
-        return getConstraint(catalog, schemaPattern, tableName, FK);
-    }
-
-    @Override
-    public List<ColumnDTO> listColumnsByTable(String catalog, String schemaPattern, String tableName) throws SQLException {
-        String sql = "select" +
-                " t.table_catalog as catalog_name," +
-                " coalesce(t.table_schema,n.nspname) as schema_name," +
-                " coalesce(t.table_name,c.relname) as table_name," +
-                " coalesce(t.column_name,attr.attname) as column_name," +
-                " coalesce(t.data_type,format_type(attr.atttypid,null)) as data_type," +
-                " t.character_maximum_length as data_length," +
-                " t.numeric_precision as data_precision," +
-                " t.numeric_scale as data_scale," +
-                " t.column_default as default_value," +
-                " coalesce(t.ordinal_position,attr.attnum) as column_ordinal," +
-                " col_description(attr.attrelid, attr.attnum) as remark," +
-                " t.is_nullable as is_nullable" +
-                " from pg_catalog.pg_class c" +
-                " join pg_catalog.pg_namespace n" +
-                " on n.oid = c.relnamespace" +
-                " left join pg_catalog.pg_attribute attr" +
-                " on attr.attrelid = c.oid" +
-                " left join information_schema.columns t" +
-                " on n.nspname = t.table_schema" +
-                " and c.relname = t.table_name" +
-                " and attr.attname = t.column_name" +
-                " where c.relkind in ('r','v','m')" +
-                " and attr.attnum > 0";
-        if (StrUtil.isNotEmpty(schemaPattern)) {
-            sql += " and coalesce(t.table_schema,n.nspname) = '" + schemaPattern + "'";
-        }
-        if (StrUtil.isNotEmpty(tableName)) {
-            sql += " and coalesce(t.table_name,c.relname) = '" + tableName + "'";
-        }
-        return super.listColumnDetails(sql);
-    }
-
-    @Override
-    public List<String> listObjectTypes() throws SQLException {
-        List<String> list = new ArrayList<String>() {{
-            add(TABLE.name());
-            add(VIEW.name());
-            add(MATERIALIZED_VIEW.name());
-            add(FOREIGN_TABLE.name());
-            add(FUNCTION.name());
-            add(INDEX.name());
-            add(SEQUENCE.name());
-        }};
-        return list.stream().map(String::toLowerCase).collect(Collectors.toList());
+    public SequenceHandler getSequenceHandler() {
+        PostgreSequenceHandler handler = new PostgreSequenceHandler();
+        handler.initialize(getDataSource(), new HashMap<>());
+        return handler;
     }
 
     @Override
     public Boolean isSupportRowEdit() {
         return true;
-    }
-
-    private List<ObjectDTO> getConstraint(String catalog, String schemaPattern, String tableName, DbObjectType type) throws SQLException {
-        List<ObjectDTO> pkList = new ArrayList<>();
-        String constraintType = "";
-        if (PK.equals(type)) {
-            constraintType = "PRIMARY KEY";
-        } else if (FK.equals(type)) {
-            constraintType = "FOREIGN KEY";
-        }
-        String sql = "select" +
-                "    t.table_catalog," +
-                "    t.table_schema," +
-                "    t.table_name," +
-                "    t.constraint_name," +
-                "    t.constraint_type" +
-                " from information_schema.table_constraints t" +
-                " where t.constraint_type = '" + constraintType + "'";
-        if (StrUtil.isNotEmpty(catalog)) {
-            sql += " and t.table_catalog ='" + catalog + "'";
-        }
-        if (StrUtil.isNotEmpty(schemaPattern)) {
-            sql += " and t.table_schema = '" + schemaPattern + "'";
-        }
-        if (StrUtil.isNotEmpty(tableName)) {
-            sql += " and t.table_name = '" + tableName + "'";
-        }
-        Connection conn = this.getConnection();
-        PreparedStatement pstm = conn.prepareStatement(sql);
-        ResultSet rs = pstm.executeQuery();
-        while (rs.next()) {
-            ObjectDTO obj = new ObjectDTO();
-            obj.setCatalogName(rs.getString("table_catalog"));
-            obj.setSchemaName(rs.getString("table_schema"));
-            obj.setObjectName(rs.getString("constraint_name"));
-            obj.setObjectType(type.name());
-            pkList.add(obj);
-        }
-        JdbcUtil.close(conn, pstm, rs);
-        return pkList;
     }
 
 
@@ -478,10 +204,4 @@ public class PostgrePluginImpl extends AbstractDataSourcePlugin {
         }
     }
 
-    @Override
-    public Map<String, TypeInfoDTO> listDataType() throws SQLException {
-        Map<String, TypeInfoDTO> map = super.listDataType();
-        map.entrySet().removeIf((entry) -> entry.getKey().startsWith("_"));
-        return map;
-    }
 }
